Move common controls to OpenKh.Tools.Common.Controls
