+++ OpenKh.Tools.Common/Controls/LayoutRendererPanel.cs
@@ -10,7 +10,7 @@ using System.Windows;
 using Xe.Drawing;
 using static kh.tools.common.DependencyPropertyUtils;
 
-namespace OpenKh.Tools.LayoutViewer.Views
+namespace OpenKh.Tools.Common.Controls
 {
     public class LayoutRendererPanel : DrawPanel
     {
