+++ OpenKh.Tools.Common/Controls/Timeline.xaml
@@ -1,9 +1,9 @@
-﻿<UserControl x:Class="OpenKh.Tools.LayoutViewer.Controls.Timeline"
+﻿<UserControl x:Class="OpenKh.Tools.Common.Controls.Timeline"
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
              xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
              xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
-             xmlns:local="clr-namespace:OpenKh.Tools.LayoutViewer.Controls"
+             xmlns:local="clr-namespace:OpenKh.Tools.Common.Controls"
              mc:Ignorable="d" 
              d:DesignHeight="120" d:DesignWidth="300"
              MouseDown="UserControl_MouseDown"
