+++ OpenKh.Tools.LayoutViewer/OpenKh.Tools.LayoutViewer.csproj
@@ -56,9 +56,6 @@
     </None>
   </ItemGroup>
   <ItemGroup>
-    <Page Update="Controls\Timeline.xaml">
-      <SubType>Designer</SubType>
-    </Page>
     <Page Update="Views\FrameView.xaml">
       <SubType>Designer</SubType>
     </Page>
