+++ OpenKh.Tools.LayoutViewer/Views/SequenceGroupView.xaml
@@ -4,6 +4,7 @@
              xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
              xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
              xmlns:local="clr-namespace:OpenKh.Tools.LayoutViewer.Views"
+             xmlns:controls="clr-namespace:OpenKh.Tools.Common.Controls;assembly=OpenKh.Tools.Common"
              mc:Ignorable="d" 
              d:DesignHeight="450" d:DesignWidth="400">
     <Grid>
@@ -60,7 +61,7 @@
                         <CheckBox IsChecked="{Binding IsVisible}"/>
                         <TextBlock Text="{Binding}"/>
                         <StackPanel.ToolTip>
-                            <local:SequenceRendererPanel
+                            <controls:SequenceRendererPanel
     Drawing="{Binding Drawing}"
     FrameIndex="{Binding FrameIndex}"
     FramesPerSecond="60"
